requirements.txt
runpod
torch
ultralytics
opencv-python-headless
Pillow
numpy
requests


